fix(#975): mount MergeFlow at App level so it survives project switch
PR #974 moved MergeFlow up to BottomPanel.tsx to survive bottom-tab
switches, but Codex round-2 flagged that BottomPanel itself only
renders when currentProject is set (App.tsx render branch). A user
switching/closing the active project mid-conflict-resolution still
unmounted MergeFlow and bypassed its close-guard.

Fix: lift the mount to App.tsx, after the {currentProject ? (...)}
conditional but inside TooltipProvider. Driven by the same
gitSlice.mergeFlowSource state. A thin AppLevelMergeFlow wrapper
isolates the gitSlice subscription so App itself does not re-render
on every modal state change.

Verified:
- npm run build clean (TS strict + Vite production)
- npx vitest run: 356/356 pass, 13 skipped (no regressions)
